+++ inst/doc/exporting_tables.R
@@ -42,81 +42,81 @@ etable(est, style.df = style.df(depvar.title = "", fixef.title = "",
 require_pander_ON()
 
 ## -----------------------------------------------------------------------------
-library(pander)
-
-etable(est, postprocess.df = pandoc.table.return, style = "rmarkdown")
-
-## -----------------------------------------------------------------------------
-my_style = style.df(depvar.title = "", fixef.title = "", 
-                    fixef.suffix = " fixed effect", yesNo = "yes")
-setFixest_etable(style.df = my_style, postprocess.df = pandoc.table.return)
+#  library(pander)
+#  
+#  etable(est, postprocess.df = pandoc.table.return, style = "rmarkdown")
 
 ## -----------------------------------------------------------------------------
-etable(est[rhs = 2], style = "rmarkdown", caption = "New default values")
+#  my_style = style.df(depvar.title = "", fixef.title = "",
+#                      fixef.suffix = " fixed effect", yesNo = "yes")
+#  setFixest_etable(style.df = my_style, postprocess.df = pandoc.table.return)
+
+## -----------------------------------------------------------------------------
+#  etable(est[rhs = 2], style = "rmarkdown", caption = "New default values")
 
 ## ---- echo = FALSE------------------------------------------------------------
-require_pander_OFF()
+#  require_pander_OFF()
 
 ## -----------------------------------------------------------------------------
-est_slopes = feols(Ozone ~ Solar.R + Wind | Day + Month[Temp], airquality)
+#  est_slopes = feols(Ozone ~ Solar.R + Wind | Day + Month[Temp], airquality)
 
 ## ---- results = 'hide'--------------------------------------------------------
-etable(est, est_slopes, tex = TRUE)
+#  etable(est, est_slopes, tex = TRUE)
 
 ## ---- include = FALSE---------------------------------------------------------
-# etable(est, est_slopes, file = "../_VIGNETTES/vignette_etable.tex", replace = TRUE)
-# etable(est, est_slopes, file = "../_VIGNETTES/vignette_etable.tex", style.tex = style.tex("aer"), fitstat = ~ r2 + n, signifCode = NA)
+#  # etable(est, est_slopes, file = "../_VIGNETTES/vignette_etable.tex", replace = TRUE)
+#  # etable(est, est_slopes, file = "../_VIGNETTES/vignette_etable.tex", style.tex = style.tex("aer"), fitstat = ~ r2 + n, signifCode = NA)
 
 ## ---- results = 'hide'--------------------------------------------------------
-etable(est, est_slopes, style.tex = style.tex("aer"), 
-       signifCode = NA, fitstat = ~ r2 + n, tex = TRUE)
+#  etable(est, est_slopes, style.tex = style.tex("aer"),
+#         signifCode = NA, fitstat = ~ r2 + n, tex = TRUE)
 
 ## -----------------------------------------------------------------------------
-set_rules = function(x, heavy, light){
-  # x: the character vector returned by etable
-  
-  tex2add = ""
-  if(!missing(heavy)){
-    tex2add = paste0("\\setlength\\heavyrulewidth{", heavy, "}\n")
-  }
-  if(!missing(light)){
-    tex2add = paste0(tex2add, "\\setlength\\lightrulewidth{", light, "}\n")
-  }
-  
-  if(nchar(tex2add) > 0){
-    x[x == "%start:tab\n"] = tex2add
-  }
-  
-  x
-}
+#  set_rules = function(x, heavy, light){
+#    # x: the character vector returned by etable
+#  
+#    tex2add = ""
+#    if(!missing(heavy)){
+#      tex2add = paste0("\\setlength\\heavyrulewidth{", heavy, "}\n")
+#    }
+#    if(!missing(light)){
+#      tex2add = paste0(tex2add, "\\setlength\\lightrulewidth{", light, "}\n")
+#    }
+#  
+#    if(nchar(tex2add) > 0){
+#      x[x == "%start:tab\n"] = tex2add
+#    }
+#  
+#    x
+#  }
 
 ## ---- results = 'hide'--------------------------------------------------------
-etable(est, est_slopes, postprocess.tex = set_rules, heavy = "0.14em", tex = TRUE)
+#  etable(est, est_slopes, postprocess.tex = set_rules, heavy = "0.14em", tex = TRUE)
 
 ## -----------------------------------------------------------------------------
-setFixest_etable(style.tex = style.tex("aer"), postprocess.tex = set_rules, 
-                 fitstat = ~ r2 + n, signifCode = NA)
+#  setFixest_etable(style.tex = style.tex("aer"), postprocess.tex = set_rules,
+#                   fitstat = ~ r2 + n, signifCode = NA)
 
 ## -----------------------------------------------------------------------------
-etable(est, est_slopes, heavy = "0.14em", tex = TRUE)
+#  etable(est, est_slopes, heavy = "0.14em", tex = TRUE)
 
 ## -----------------------------------------------------------------------------
-fitstat_register(type = "p_s", alias = "pvalue (standard)",
-                 fun = function(x) pvalue(x, se = "s")["Solar.R"])
-
-fitstat_register(type = "p_h", alias = "pvalue (Heterosk.)",
-                 fun = function(x) pvalue(x, se = "h")["Solar.R"])
-
-fitstat_register(type = "p_day", alias = "pvalue (Day)",
-                 fun = function(x) pvalue(x, cluster = "Day")["Solar.R"])
-
-fitstat_register(type = "p_month", alias = "pvalue (Month)",
-                 fun = function(x) pvalue(x, cluster = "Month")["Solar.R"])
-
-# We first reset the default values set in the previous sections
-setFixest_etable(reset = TRUE)
-# Now we display the results with the new fit statistics
-etable(est, fitstat = ~ . + p_s + p_h + p_day + p_month)
+#  fitstat_register(type = "p_s", alias = "pvalue (standard)",
+#                   fun = function(x) pvalue(x, se = "s")["Solar.R"])
+#  
+#  fitstat_register(type = "p_h", alias = "pvalue (Heterosk.)",
+#                   fun = function(x) pvalue(x, se = "h")["Solar.R"])
+#  
+#  fitstat_register(type = "p_day", alias = "pvalue (Day)",
+#                   fun = function(x) pvalue(x, cluster = "Day")["Solar.R"])
+#  
+#  fitstat_register(type = "p_month", alias = "pvalue (Month)",
+#                   fun = function(x) pvalue(x, cluster = "Month")["Solar.R"])
+#  
+#  # We first reset the default values set in the previous sections
+#  setFixest_etable(reset = TRUE)
+#  # Now we display the results with the new fit statistics
+#  etable(est, fitstat = ~ . + p_s + p_h + p_day + p_month)
 
 ## ---- eval = FALSE------------------------------------------------------------
 #  summary(.l(est, est_slopes), cluster = ~ Month)